requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pytubefix==5.1.1
cohere==4.25.0
aiohttp==3.9.1
deepgram-sdk==3.2.7
python-dotenv==0.20.0
gunicorn==20.1.0
flask-cors==4.0.0
typing-extensions
google-api-python-client==2.108.0